Add extraTile, better player collisions
Added an extra tile that will evenually be pushed into the grid but currently floats in the margins.

Added better player collisions. It finds the closest block after a collision and moves as close to that as possible. TODO get all collision blocks and get as close as possible to the closest from starting position.